Make AccountLogo use the Socialtext::Avatar role

+++ nlw/lib/Socialtext/Avatar.pm
@@ -10,31 +10,20 @@ use Socialtext::SQL qw(:txn :exec get_dbh);
 use Socialtext::SQL::Builder qw(sql_insert sql_update);
 
 sub DefaultPhoto {
-    my $self_or_class = shift;
-    my $size = shift || die 'size required';
+    my $class = shift;
+    my $version = shift || die 'version required';
 
     # get the path to the image, on *disk*
-    my $skin = Socialtext::Skin->new( name => 'common' );
+    my $skin = Socialtext::Skin->new( name => $class->default_skin );
     my $dir = File::Spec->catfile($skin->skin_path, "images");
 
-    my $default_arg = "default_$size";
-    my $file = $self_or_class->$default_arg || die "no $default_arg!";
+    my $default_arg = "default_$version";
+    my $file = $class->$default_arg || die "no $default_arg!";
     my $blob = Socialtext::File::get_contents_binary("$dir/$file");
     return \$blob;
 }
 
-my %SIZES = (
-    small => {
-        width => 27,
-        column => 'small_photo_image',
-    },
-    large => {
-        width => 62,
-        column => 'photo_image',
-    },
-);
-
-requires qw(cache table id_column id synonyms default_large default_small);
+requires qw(cache table versions id_column id resize default_skin);
 
 has 'cache_dir' => (
     is => 'ro', isa => 'Str',
@@ -49,41 +38,26 @@ sub _build_cache_dir {
     return $cache_dir;
 }
 
-# Blobs:
-has 'large' => (
-    is => 'rw', isa => 'ScalarRef',
-    lazy_build => 1,
-);
-sub _build_large { $_[0]->_load('large') }
-
-has 'small' => (
-    is => 'rw', isa => 'ScalarRef',
-    lazy_build => 1,
-);
-sub _build_small { $_[0]->_load('small') }
-
 sub set {
     my $self = shift;
     my $blob_ref = shift;
 
     eval {
-        for my $size (qw(small large)) {
-            my $width = $SIZES{$size}{width} || die "Unknown size: $size";
+        my %blobs;
+        for my $version ($self->versions) {
             my ($fh, $filename) = tempfile;
             print $fh $$blob_ref;
             close $fh or die "Invalid image: $!";
 
-            Socialtext::Image::extract_rectangle(
-                image_filename => $filename,
-                width => $width,
-                height => $width,
-            );
+            $self->resize($version, $filename);
+
             my $contents = Socialtext::File::get_contents_binary($filename);
-            $self->$size(\$contents);
+            $self->$version(\$contents);
+            $blobs{$version} = \$contents;
         }
 
-        $self->_save_db(small => $self->small, large => $self->large);
-        $self->_save_cache(small => $self->small, large => $self->large);
+        $self->_save_db(%blobs);
+        $self->_save_cache(%blobs);
     };
     # check if there were any problems with the image format
     if ($@) {
@@ -106,19 +80,19 @@ sub purge {
     # Remove from cache
     my $cache_dir = $self->cache_dir;
     my $lock_fh = Socialtext::File::write_lock("$cache_dir/.lock");
-    for my $size (qw(small large)) {
-        unlink "$cache_dir/$id-$size.png";
-        my @symlinks = map { s#/#\%2f#g; "$cache_dir/$_-$size.png" }
-                       $self->synonyms;
-        unlink @symlinks;
+    for my $version ($self->versions) {
+        unlink "$cache_dir/$id-$version.png";
+        if ($self->can('synonyms')) {
+            my @symlinks = map { s#/#\%2f#g; "$cache_dir/$_-$version.png" }
+                           $self->synonyms;
+            unlink @symlinks;
+        }
     }
 }
 
 
 sub _save_db {
     my ($self, %blobs) = @_;
-    die "Must call save with both small and large blobs!"
-        unless $blobs{small} and $blobs{large};
 
     my $dbh = get_dbh;
 
@@ -136,22 +110,23 @@ sub _save_db {
 
         my $sth;
         if ($exists) {
-            $sth = $dbh->prepare("
-                UPDATE $table
-                   SET photo_image = ?,
-                       small_photo_image = ?
-                 WHERE $id_column = ?
-            ");
+            my $sets = join ", ", map { "$_ = ?" } $self->versions;
+            $sth = $dbh->prepare(
+                "UPDATE $table SET $sets WHERE $id_column = ?"
+            );
         }
         else {
+            my $cols = join ", ", $self->versions;
             $sth = $dbh->prepare("
-                INSERT INTO $table (photo_image, small_photo_image, $id_column)
-                VALUES (?,?,?)
+                INSERT INTO $table ($cols, $id_column) VALUES (?,?,?)
             ");
         }
-        $sth->bind_param(1, ${$blobs{large}}, {pg_type => PG_BYTEA});
-        $sth->bind_param(2, ${$blobs{small}}, {pg_type => PG_BYTEA});
-        $sth->bind_param(3, $self->id);
+
+        my $n = 1;
+        for my $version ($self->versions) {
+            $sth->bind_param($n++, ${$blobs{$version}}, {pg_type => PG_BYTEA});
+        }
+        $sth->bind_param($n++, $self->id);
         $sth->execute;
 
         die "unable to update image" unless ($sth->rows == 1);
@@ -159,8 +134,8 @@ sub _save_db {
         sql_commit($dbh) unless $txn;
     };
     if ($@) {
-        sql_rollback($dbh) unless $txn;
         warn $@;
+        sql_rollback($dbh) unless $txn;
         return "SQL Error";
     }
 }
@@ -173,42 +148,38 @@ sub _save_cache {
 
     my $lock_fh = Socialtext::File::write_lock("$cache_dir/.lock");
 
-    for my $size (keys %blobs) {
-        my $file = "$cache_dir/$id-$size.png";
+    for my $version (keys %blobs) {
+        my $file = "$cache_dir/$id-$version.png";
         my $temp = "$file.tmp";
-        my @symlinks = map { s#/#\%2f#g; "$cache_dir/$_-$size.png" }
-                       $self->synonyms;
-
-        eval {
-            Socialtext::File::set_contents_binary($temp, $blobs{$size});
-            rename $temp, $file;
+        Socialtext::File::set_contents_binary($temp, $blobs{$version});
+        rename $temp, $file;
+        if ($self->can('synonyms')) {
+            my @symlinks = map { s#/#\%2f#g; "$cache_dir/$_-$version.png" }
+                           $self->synonyms;
             for my $link (@symlinks) {
                 unlink $link; # fail = ok
                 link $file => $link;
             }
-        };
-        warn $@ if $@;
+        }
     }
 }
 
-sub _load {
-    my $self = shift;
-    my $size = shift;
+sub load {
+    my ($self, $version) = @_;
 
-    my $column = $SIZES{$size}{column} || die "no known column for $size";
     my $table = $self->table;
     my $id_column = $self->id_column;
 
     my $sth = sql_execute(
-        "SELECT $column FROM $table WHERE $id_column = ?", $self->id
+        "SELECT $version FROM $table WHERE $id_column = ?", $self->id
     );
     my $blob;
     $sth->bind_columns(\$blob);
     $sth->fetch();
     $sth->finish();
 
-    my $blob_ref = $blob ? \$blob : $self->DefaultPhoto($size);
-    $self->_save_cache($size => $blob_ref);
+    my $blob_ref = $blob ? \$blob : $self->DefaultPhoto($version);
+    $self->_save_cache($version => $blob_ref);
     return $blob_ref;
 }
